Deal with curses not drawing bottom-right char
Curses can't draw the bottom-right character of a window, because it
tries to wrap the cursor and cannot (because it's out of lines)! It
*can*, however, draw the bottom-right visible corner of a pad, provided
the cursor can wrap to a (maybe invisible) blank line.

So here, we'll just make the map a pad that's one line too tall, and
never draw the last line.