Shuffle a bunch of code around
The internet tells me that Python doesn't really stick with
one-class-per-file in the way Perl usually does, and they prefer flat
over deep hierarchies. I suspect that's because Python's import
machinery is...not great, but fine.

This is all just a prelude to defining map objects from a file and not
having to write a bunch of dumb classes.